refactor: remove dead AI surface code
Remove the micro-explanations feature track (confirmed zero runtime callers),
legacy schema.ts, and dead re-exports/types across the AI module.

Changes:
- Delete micro-explanations.ts and its test, prompt builders, auth config types, router config method, and env config
- Delete schema.ts (TaskSchema/ModuleSchema/PlanSchema/PlanOutput) and its test — only used by tests, not production
- Remove getGenerationProvider re-export from provider.ts backward-compat shim
- Remove PDF_SECTION_CONTENT_LIMIT re-export from prompts.ts
- Remove StreamingEventValidated unused type from streaming/schema.ts
- Remove ParsedGeneration/ParsedModule re-exports from orchestrator.ts, update regeneration-worker to import from parser.ts
- Consolidate timeout constants: export DEFAULT_GENERATION_TIMEOUT_MS and DEFAULT_GENERATION_EXTENSION_MS from timeout.ts, import in openrouter.ts
- Update test imports (mock.spec.ts, prompts.spec.ts, fixtures/prompts.ts)
- Update AI AGENTS.md to remove schema.ts reference
