+++ src/lib/ai/provider.ts
@@ -70,6 +70,3 @@ export class ProviderInvalidResponseError extends ProviderError {
     this.name = 'ProviderInvalidResponseError';
   }
 }
-
-// Re-export factory function from provider-factory module
-export { getGenerationProvider } from '@/lib/ai/provider-factory';
