+++ tests/fixtures/prompts.ts
@@ -1,7 +1,4 @@
-import type {
-  MicroExplanationPromptParams,
-  PromptParams,
-} from '@/lib/ai/prompts';
+import type { PromptParams } from '@/lib/ai/prompts';
 
 const DEFAULT_PROMPT_PARAMS: PromptParams = {
   topic: 'TypeScript',
@@ -10,26 +7,11 @@ const DEFAULT_PROMPT_PARAMS: PromptParams = {
   weeklyHours: 10,
 };
 
-const DEFAULT_MICRO_EXPLANATION_PARAMS: MicroExplanationPromptParams = {
-  topic: 'React Hooks',
-  taskTitle: 'Understanding useState',
-  skillLevel: 'beginner',
-};
-
 export function createPromptParams(
   overrides: Partial<PromptParams> = {}
 ): PromptParams {
   return {
     ...DEFAULT_PROMPT_PARAMS,
-    ...overrides,
-  };
-}
-
-export function createMicroExplanationParams(
-  overrides: Partial<MicroExplanationPromptParams> = {}
-): MicroExplanationPromptParams {
-  return {
-    ...DEFAULT_MICRO_EXPLANATION_PARAMS,
     ...overrides,
   };
 }
